Move util files into utils
