+++ src/constants/index.ts
@@ -1,2 +1,4 @@
 export const WPOKT_ADDRESS = '0x2f363dd061cc8b3411c3c91c0cfac0fa1b62f656';
 export const TOKEN_GEYSER_ADDRESS = '0x746218704841983de2ca941dd91598e68c369025';
+export const WPOKT_GENESIS_FARM = '0x746218704841983de2ca941dd91598e68c369025';
+export const WPOKT_SUBGRAPH_URL = process.env.REACT_APP_FARM_SUBGRAPH_URL;
