--- src/util/constants.ts
@@ -1,3 +1,0 @@
-export const WPOKT_GENESIS_FARM = '0x746218704841983de2ca941dd91598e68c369025';
-
-export const WPOKT_SUBGRAPH_URL = process.env.REACT_APP_FARM_SUBGRAPH_URL;
